backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM
google-genai>=1.0.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Async
httpx==0.26.0
greenlet>=3.0.0

# Scheduling (for later)
celery==5.3.4
redis==5.0.1

# Utilities
tenacity==8.2.3
